Add TangentialPotential
