+++ src/ipc/friction/smooth_friction_mollifier.cpp
@@ -11,7 +11,7 @@ double f0_SF(const double s, const double epsv)
     if (std::abs(s) >= epsv) {
         return s;
     }
-    return s * s * (-s / (3 * epsv) + 1) / epsv + epsv / 3;
+    return s * s * (1 - s / (3 * epsv)) / epsv + epsv / 3;
 }
 
 double f1_SF_over_x(const double s, const double epsv)
@@ -20,10 +20,10 @@ double f1_SF_over_x(const double s, const double epsv)
     if (std::abs(s) >= epsv) {
         return 1 / s;
     }
-    return (-s / epsv + 2) / epsv;
+    return (2 - s / epsv) / epsv;
 }
 
-double df1_x_minus_f1_over_x3(const double s, const double epsv)
+double df1_SF_x_minus_f1_SF_over_x3(const double s, const double epsv)
 {
     assert(epsv > 0);
     if (std::abs(s) >= epsv) {
